chore: fix pr comments
